requirements.txt
affine==2.2.2
arrow==0.13.1
attrs==18.2.0
backcall==0.1.0
boto3==1.9.92
botocore==1.12.92
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
cycler==0.10.0
dateparser==0.7.1
decorator==4.3.2
Django==2.1.7
django-extensions==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
docutils==0.14
Fiona==1.8.4
geopandas==0.4.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==3.0.2
munch==2.3.2
numpy==1.16.1
pandas==0.24.1
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
pyproj==1.9.6
python-dateutil==2.8.0
pytz==2018.9
rasterio==1.0.18
regex==2019.2.21
requests==2.21.0
s3transfer==0.2.0
Shapely==1.6.4.post2
six==1.12.0
snuggs==1.4.2
times==0.7
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24.1
wcwidth==0.1.7
